# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Luxe Clothes Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Settings {
  id: number;
  storeName: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  address?: string | null;
  deliveryFee: number;
  /** @nullable */
  freeDeliveryThreshold?: number | null;
  /** @nullable */
  heroImageUrl?: string | null;
  /** @nullable */
  aboutText?: string | null;
}

export interface SettingsUpdate {
  storeName?: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  address?: string | null;
  deliveryFee?: number;
  /** @nullable */
  freeDeliveryThreshold?: number | null;
  /** @nullable */
  heroImageUrl?: string | null;
  /** @nullable */
  aboutText?: string | null;
}

export interface Category {
  id: number;
  name: string;
  /** @nullable */
  imageUrl?: string | null;
  productCount: number;
}

export interface CategoryAdmin {
  id: number;
  name: string;
  /** @nullable */
  imageUrl?: string | null;
  productCount: number;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  stock: number;
  sizes: string[];
  colors: string[];
  images: string[];
  featured: boolean;
  createdAt: string;
}

export interface CategoryWithProducts {
  id: number;
  name: string;
  /** @nullable */
  imageUrl?: string | null;
  products: Product[];
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface ProductList {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface AdminProductList {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @minimum 0 */
  price: number;
  /** @nullable */
  salePrice?: number | null;
  /** @minimum 0 */
  stock: number;
  sizes: string[];
  colors: string[];
  images: string[];
  featured?: boolean;
}

export interface ProductUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  categoryId?: number | null;
  price?: number;
  /** @nullable */
  salePrice?: number | null;
  stock?: number;
  sizes?: string[];
  colors?: string[];
  images?: string[];
  featured?: boolean;
}

export type CouponDiscountType = typeof CouponDiscountType[keyof typeof CouponDiscountType];


export const CouponDiscountType = {
  percentage: 'percentage',
  fixed: 'fixed',
} as const;

export interface Coupon {
  id: number;
  code: string;
  discountType: CouponDiscountType;
  discountValue: number;
  active: boolean;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
}

export type CouponInputDiscountType = typeof CouponInputDiscountType[keyof typeof CouponInputDiscountType];


export const CouponInputDiscountType = {
  percentage: 'percentage',
  fixed: 'fixed',
} as const;

export interface CouponInput {
  /** @minLength 1 */
  code: string;
  discountType: CouponInputDiscountType;
  /** @minimum 0 */
  discountValue: number;
  active?: boolean;
  /** @nullable */
  expiresAt?: string | null;
}

export interface CouponValidateInput {
  code: string;
  subtotal: number;
}

export type CouponValidateResultDiscountType = typeof CouponValidateResultDiscountType[keyof typeof CouponValidateResultDiscountType];


export const CouponValidateResultDiscountType = {
  percentage: 'percentage',
  fixed: 'fixed',
} as const;

export interface CouponValidateResult {
  couponId: number;
  code: string;
  discountType: CouponValidateResultDiscountType;
  discountValue: number;
  discountAmount: number;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  /** @nullable */
  productImage?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
}

export interface OrderItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  processing: 'processing',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  orderNumber: string;
  customerName: string;
  phone: string;
  address: string;
  city: string;
  /** @nullable */
  notes?: string | null;
  subtotal: number;
  deliveryFee: number;
  discount: number;
  total: number;
  status: OrderStatus;
  /** @nullable */
  couponCode?: string | null;
  items: OrderItem[];
  createdAt: string;
}

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  address: string;
  /** @minLength 1 */
  city: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  couponCode?: string | null;
  /** @minItems 1 */
  items: OrderItemInput[];
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  processing: 'processing',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface AdminOrderList {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface Admin {
  id: number;
  username: string;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminLoginResult {
  token: string;
  admin: Admin;
}

export interface TopProduct {
  productId: number;
  productName: string;
  /** @nullable */
  productImage?: string | null;
  totalSold: number;
  revenue: number;
}

export interface DashboardStats {
  totalOrders: number;
  totalRevenue: number;
  totalProducts: number;
  pendingOrders: number;
  processingOrders: number;
  deliveredOrders: number;
  cancelledOrders: number;
  recentOrders: Order[];
  topProducts: TopProduct[];
}

export interface UploadResult {
  url: string;
}

export type ListProductsParams = {
categoryId?: number;
search?: string;
sort?: ListProductsSort;
featured?: boolean;
page?: number;
limit?: number;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  newest: 'newest',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
} as const;

export type AdminListProductsParams = {
page?: number;
limit?: number;
search?: string;
categoryId?: number;
};

export type AdminListOrdersParams = {
page?: number;
limit?: number;
status?: AdminListOrdersStatus;
search?: string;
};

export type AdminListOrdersStatus = typeof AdminListOrdersStatus[keyof typeof AdminListOrdersStatus];


export const AdminListOrdersStatus = {
  processing: 'processing',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;