# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Luxe Clothes Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a single category with its products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category with products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryWithProducts"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Products (public) ────────────────────────────────────────────
...
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code and return discount info
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidateInput"
      responses:
        "200":
          description: Valid coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidateResult"
        "400":
          description: Invalid or expired coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Orders (guest) ───────────────────────────────────────────────
...
  /admin/auth/login:
    post:
      operationId: adminLogin
      tags: [admin-auth]
      summary: Admin login — returns JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/auth/me:
    get:
      operationId: adminMe
      tags: [admin-auth]
      summary: Get current admin info (requires JWT)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Admin Dashboard ────────────────────────────────────────────────
...
  /admin/categories:
    get:
      operationId: adminListCategories
      tags: [admin-categories]
      summary: Admin list all categories with product counts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Categories with counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryAdmin"
    post:
      operationId: adminCreateCategory
      tags: [admin-categories]
      summary: Create a category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/categories/{id}:
    put:
      operationId: adminUpdateCategory
      tags: [admin-categories]
      summary: Update a category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteCategory
      tags: [admin-categories]
      summary: Delete a category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Admin Orders ──────────────────────────────────────────────────
...
  /admin/coupons:
    get:
      operationId: adminListCoupons
      tags: [admin-coupons]
      summary: List all coupons
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: adminCreateCoupon
      tags: [admin-coupons]
      summary: Create a coupon
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/coupons/{id}:
    delete:
      operationId: adminDeleteCoupon
      tags: [admin-coupons]
      summary: Delete a coupon
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Admin Settings ────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Settings ─────────────────────────────────────────────────────
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        imageUrl:
          type: ["string", "null"]
        productCount:
          type: integer
      required: [id, name, productCount]
    CategoryAdmin:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        imageUrl:
          type: ["string", "null"]
        productCount:
          type: integer
      required: [id, name, productCount]
    CategoryWithProducts:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        imageUrl:
          type: ["string", "null"]
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
      required: [id, name, products]
    CategoryInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        imageUrl:
          type: ["string", "null"]
      required: [name]
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        imageUrl:
          type: ["string", "null"]

    # ── Product ───────────────────────────────────────────────────────
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        stock:
          type: integer
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        featured:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, price, stock, sizes, colors, images, featured, createdAt]
...
    Coupon:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        active:
          type: boolean
        expiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, code, discountType, discountValue, active, createdAt]
    CouponInput:
      type: object
      properties:
        code:
          type: string
          minLength: 1
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
          minimum: 0
        active:
          type: boolean
          default: true
        expiresAt:
          type: ["string", "null"]
      required: [code, discountType, discountValue]
    CouponValidateInput:
      type: object
      properties:
        code:
          type: string
        subtotal:
          type: number
      required: [code, subtotal]
    CouponValidateResult:
      type: object
      properties:
        couponId:
          type: integer
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        discountAmount:
          type: number
      required: [couponId, code, discountType, discountValue, discountAmount]

    # ── Order ─────────────────────────────────────────────────────────
...
    Admin:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
      required: [id, username]
    AdminLoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AdminLoginResult:
      type: object
      properties:
        token:
          type: string
        admin:
          $ref: "#/components/schemas/Admin"
      required: [token, admin]

    # ── Dashboard ─────────────────────────────────────────────────────
...